refactor: Organize types into modules
